+++ src/generate-with-init.ts
@@ -23,11 +23,7 @@ export function generateWithInit(
 
   for (const sourceUnit of transform.asts()) {
     for (const contract of findAll('ContractDefinition', sourceUnit)) {
-      if (
-        contract.contractKind !== 'contract' ||
-        contract.abstract ||
-        hasConstructorOverride(contract)
-      ) {
+      if (contract.contractKind !== 'contract' || contract.abstract) {
         continue;
       }
 
@@ -51,6 +47,14 @@ export function generateWithInit(
 
       const renamedContract = renameContract(contract.name);
 
+      let parents: string[] = [];
+      let statements: string[] = [];
+      if (hasConstructorOverride(contract)) {
+        parents = [`${renameContract(contract.name)}(${argNames})`];
+      } else {
+        statements = [`__${contract.name}_init(${argNames});`];
+      }
+
       res.push(
         `import "${relativePath(path.dirname(destPath), renamePath(sourceUnit.absolutePath))}";`,
         ``,
@@ -58,10 +62,11 @@ export function generateWithInit(
         [
           [
             `constructor(${argsList})`,
+            ...parents,
             ...(satisfies(pragmaVersion, '>=0.7') ? [] : [`public`]),
             `payable initializer {`,
           ].join(' '),
-          [`__${contract.name}_init(${argNames});`],
+          statements,
           `}`,
         ],
         `}`,