+++ src/transform-0.8.test.ts
@@ -13,6 +13,8 @@ import {
   transformConstructor,
 } from './transformations/transform-constructor';
 import { renameInheritdoc } from './transformations/rename-inheritdoc';
+import { addStorageGaps } from './transformations/add-storage-gaps';
+import { renameContractDefinition } from './transformations/rename-contract-definition';
 
 const test = _test as TestFn<Context>;
 
@@ -43,6 +45,7 @@ test('rename parents in solidity 0.8', t => {
 test('correctly index when utf8 characters', t => {
   const file = 'contracts/TransformUtf8Chars.sol';
   t.context.transform.apply(renameIdentifiers);
+  t.context.transform.apply(renameContractDefinition);
   t.snapshot(t.context.transform.results()[file]);
 });
 
@@ -52,5 +55,15 @@ test('preserves immutable if allowed', t => {
   t.context.transform.apply(removeLeftoverConstructorHead);
   t.context.transform.apply(removeStateVarInits);
   t.context.transform.apply(removeImmutable);
+  t.snapshot(t.context.transform.results()[file]);
+});
+
+test('custom contract size', t => {
+  const file = 'contracts/TransformCustomSize.sol';
+  t.context.transform.apply(transformConstructor());
+  t.context.transform.apply(removeLeftoverConstructorHead);
+  t.context.transform.apply(removeStateVarInits);
+  t.context.transform.apply(removeImmutable);
+  t.context.transform.apply(addStorageGaps);
   t.snapshot(t.context.transform.results()[file]);
 });