+++ src/transformations/add-public-getters.ts
@@ -1,23 +1,17 @@
-import {ElementaryTypeName, SourceUnit, TypeName, UserDefinedTypeName, VariableDeclaration } from 'solidity-ast';
+import {SourceUnit, TypeName } from 'solidity-ast';
 
-import { getConstructor, getNodeBounds } from '../solc/ast-utils';
-import { Transformation, TransformHelper } from './type';
-import { buildSuperCallsForChain2 } from './utils/build-super-calls-for-chain';
+import { Transformation } from './type';
 import { findAll } from 'solidity-ast/utils';
 import { TransformerTools } from '../transform';
 import { newFunctionPosition } from './utils/new-function-position';
-import { formatLines } from './utils/format-lines';
-import { hasConstructorOverride, hasOverride } from '../utils/upgrades-overrides';
+import { hasOverride } from '../utils/upgrades-overrides';
 import { getVarStorageName } from './utils/get-var-storage-name';
 import { renameContract, renamePath } from '../rename';
-import { Node } from 'solidity-ast/node';
-import path from 'path';
 
 export function* addPublicGetters(
   sourceUnit: SourceUnit,
   tools: TransformerTools,
 ): Generator<Transformation> {
-
   const { resolver } = tools;
 
   for (const contractNode of findAll('ContractDefinition', sourceUnit)) {
@@ -25,14 +19,13 @@ export function* addPublicGetters(
       continue;
     }
 
-    const {name} = contractNode;
-
     const publicVariables = [...findAll('VariableDeclaration', contractNode)].filter(
-        v =>
-            v.stateVariable && (v.visibility === "public") &&
-            !v.constant &&
-            !hasOverride(v, 'state-variable-assignment') &&
-            v.functionSelector,
+      v =>
+        v.stateVariable &&
+        v.visibility === 'public' &&
+        !v.constant &&
+        !hasOverride(v, 'state-variable-assignment') &&
+        v.functionSelector,
     );
 
     if (publicVariables.length > 0) {
@@ -40,22 +33,21 @@ export function* addPublicGetters(
       let getterFuncsText = '';
 
       for (const varDecl of publicVariables) {
-
         const argumentsText: string[] = [];
         // VariableDeclaration nodes for function parameters or state variables will always
         // have their typeName fields defined.
         let nextType = varDecl.typeName!;
         let returnType;
         while (true) {
-          if (nextType.nodeType === "Mapping") {
+          if (nextType.nodeType === 'Mapping') {
             const canonicalType = canonicalAbiTypeForElementaryOrUserDefinedTypes(
-                nextType.keyType
+              nextType.keyType,
             )!;
             argumentsText.push(`${canonicalType} arg${argumentsText.length.toString()}`);
 
             nextType = nextType.valueType;
           } else {
-            if (nextType.nodeType === "ArrayTypeName") {
+            if (nextType.nodeType === 'ArrayTypeName') {
               argumentsText.push(`uint256 arg${argumentsText.length.toString()}`);
               nextType = nextType.baseType;
             }
@@ -63,7 +55,7 @@ export function* addPublicGetters(
             if (nextType.nodeType === 'UserDefinedTypeName') {
               const userVarDecl = resolver.resolveScope(nextType.referencedDeclaration)!.node;
               returnType = nextType.pathNode!.name;
-              if (['ArrayDefinition','StructDefinition'].indexOf(userVarDecl.nodeType) > -1) {
+              if (['ArrayDefinition', 'StructDefinition'].indexOf(userVarDecl.nodeType) > -1) {
                 if (('scope' in userVarDecl) && (userVarDecl.scope !== contractNode.id)) {
                   returnType = renamePath(returnType);
                 }
@@ -106,24 +98,24 @@ export function* addPublicGetters(
 
 function isContractType(param: any) {
   return (
-      (param.typeName?.nodeType === "UserDefinedTypeName" ||
-          param?.nodeType === "UserDefinedTypeName") &&
-      param.typeDescriptions?.typeString !== undefined &&
-      param.typeDescriptions.typeString.startsWith("contract ")
+    (param.typeName?.nodeType === 'UserDefinedTypeName' ||
+      param?.nodeType === 'UserDefinedTypeName') &&
+    param.typeDescriptions?.typeString !== undefined &&
+    param.typeDescriptions.typeString.startsWith('contract ')
   );
 }
 
 function isEnumType(param: any) {
   return (
-      (param.typeName?.nodeType === "UserDefinedTypeName" ||
-          param?.nodeType === "UserDefinedTypeName") &&
-      param.typeDescriptions?.typeString !== undefined &&
-      param.typeDescriptions.typeString.startsWith("enum ")
+    (param.typeName?.nodeType === 'UserDefinedTypeName' ||
+      param?.nodeType === 'UserDefinedTypeName') &&
+    param.typeDescriptions?.typeString !== undefined &&
+    param.typeDescriptions.typeString.startsWith("enum ")
   );
 }
 
 function canonicalAbiTypeForElementaryOrUserDefinedTypes(typeName: TypeName): string | undefined {
-  if (typeName.nodeType === "ElementaryTypeName") {
+  if (typeName.nodeType === 'ElementaryTypeName') {
     return toCanonicalAbiType(typeName.name);
   }
 
@@ -139,36 +131,36 @@ function canonicalAbiTypeForElementaryOrUserDefinedTypes(typeName: TypeName): st
 }
 
 function toCanonicalAbiType(type: string): string {
-  if (type.startsWith("int[")) {
+  if (type.startsWith('int[')) {
     return `int256${type.slice(3)}`;
   }
 
-  if (type === "int") {
-    return "int256";
+  if (type === 'int') {
+    return 'int256';
   }
 
-  if (type.startsWith("uint[")) {
+  if (type.startsWith('uint[')) {
     return `uint256${type.slice(4)}`;
   }
 
-  if (type === "uint") {
-    return "uint256";
+  if (type === 'uint') {
+    return 'uint256';
   }
 
-  if (type.startsWith("fixed[")) {
+  if (type.startsWith('fixed[')) {
     return `fixed128x128${type.slice(5)}`;
   }
 
-  if (type === "fixed") {
-    return "fixed128x128";
+  if (type === 'fixed') {
+    return 'fixed128x128';
   }
 
-  if (type.startsWith("ufixed[")) {
+  if (type.startsWith('ufixed[')) {
     return `ufixed128x128${type.slice(6)}`;
   }
 
-  if (type === "ufixed") {
-    return "ufixed128x128";
+  if (type === 'ufixed') {
+    return 'ufixed128x128';
   }
 
   return type;
@@ -182,5 +174,3 @@ function renamePathNode(pathNode: string) : string {
 
   return pathNode;
 }
-
-